Contact page updates and added a todo file
I added content to the contact page mainly, i was able to add the contact form, not fully setup yet, which is logged in the TODO text file which i also added. which will help me keep track of projects. this principle of logging what i need to do inside the folder/outside where it is secured allows me to be able to ensure i klnow what to do next.